rename get_graphiql_html.py to graphiql.py
